bot2/requirements.txt
faiss-cpu
fastapi
httpx
python-dotenv
sentence-transformers
scikit-learn
uvicorn
tqdm
numpy
scipy
regex
requests
pillow
